Runned Checkstyle and meet all requirements
