+++ src/view/GamePanel.java
@@ -1,17 +1,17 @@
 package view;
 
 import Model.GamePackage.Player;
-import view.boardSpaces.CardSpace;
-import view.boardSpaces.FreeParkingSpace;
-import view.boardSpaces.GoSpace;
-import view.boardSpaces.GoToJailSpace;
-import view.boardSpaces.JailSpace;
-import view.boardSpaces.PropertySpace;
-import view.boardSpaces.PropertySpace.Position;
-import view.boardSpaces.RailRoadSpace;
-import view.boardSpaces.Spaces;
-import view.boardSpaces.TaxSpace;
-import view.boardSpaces.UtilitiesSpace;
+import view.boardspaces.CardSpace;
+import view.boardspaces.FreeParkingSpace;
+import view.boardspaces.GoSpace;
+import view.boardspaces.GoToJailSpace;
+import view.boardspaces.JailSpace;
+import view.boardspaces.PropertySpace;
+import view.boardspaces.PropertySpace.Position;
+import view.boardspaces.RailRoadSpace;
+import view.boardspaces.Spaces;
+import view.boardspaces.TaxSpace;
+import view.boardspaces.UtilitiesSpace;
 import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.GridBagConstraints;
