+++ src/view/boardspaces/CardSpace.java
@@ -1,4 +1,4 @@
-package view.boardSpaces;
+package view.boardspaces;
 
 import java.awt.Color;
 import java.awt.Dimension;
