+++ src/view/boardspaces/JailSpace.java
@@ -1,4 +1,4 @@
-package view.boardSpaces;
+package view.boardspaces;
 
 import java.awt.*;
 import java.io.File;
@@ -8,13 +8,13 @@ import javax.swing.*;
 import javax.swing.border.Border;
 
 /**********************************************************************
- * The GoSpace class creates a JPanel that resembles a Go Square from
- * the game of Monopoly.
+ * The JailSpace class creates a JPanel that resembles a Jail
+ * Square from the game of Monopoly.
  *
- * @author David Baas
+ * @author Dustin Ecker
  * @version 3/26/2018
  *********************************************************************/
-public class GoSpace extends Spaces {
+public class JailSpace extends Spaces {
 
   /**
    * A static int constant to track the length of an edge in the square.
@@ -22,40 +22,39 @@ public class GoSpace extends Spaces {
   private static final int LENGTH = 100;
 
   /**
-   * An image to hold the image of the Go icon.
+   * An image to hold the image of the Jail icon.
    */
-  private static Image goImg;
-
+  private static Image jailImg;
 
   /********************************************************************
    * The constructor initializes the image.
    *
    * @throws IOException If there is an error in loading the image.
    *******************************************************************/
-  public GoSpace() throws IOException {
+  public JailSpace() throws IOException {
     super(true);
 
-    goImg = ImageIO.read(new File("res/goImg.gif"));
+    jailImg = ImageIO.read(new File("res/jailImg.png"));
 
     Border blackLine = BorderFactory.createLineBorder(Color.BLACK);
     setBorder(blackLine);
   }
 
-  /*******************************************************************
-   * The paintComponent draws the Go Image onto the JPanel.
+  /********************************************************************
+   * The paintComponent draws Jail image onto the JPanel.
    *
    * @param g Graphics component of this panel.
-   ******************************************************************/
+   ********************************************************************/
   public void paintComponent(Graphics g) {
     super.paintComponent(g);
-    g.drawImage(goImg, 0, 0, LENGTH, LENGTH, null);
+    g.drawImage(jailImg, 0, 0, LENGTH, LENGTH, null);
     drawTokens(g);
   }
 
   /********************************************************************
    * The getPreferredSize method is used to Lock the size of the Panel
    * to the correct size.
-   * @return The dimensions of the GoSquare.
+   * @return The dimensions of the JailSquare.
    *******************************************************************/
   @Override
   public Dimension getPreferredSize() {
@@ -65,7 +64,7 @@ public class GoSpace extends Spaces {
   /********************************************************************
    * The getMinimumSize method refers to the getPreferredSize method
    * to lock the size of the panel.
-   * @return getPreferredSize the dimensions of the GoSquare.
+   * @return getPreferredSize the dimensions of the JailSquare.
    *******************************************************************/
   @Override
   public Dimension getMinimumSize() {
@@ -75,7 +74,7 @@ public class GoSpace extends Spaces {
   /********************************************************************
    * The getMaximumSize method refers to the getPreferredSize method
    * to lock the size of the panel.
-   * @return getPreferredSize the dimensions of the GoSquare.
+   * @return getPreferredSize the dimensions of the JailSquare.
    *******************************************************************/
   @Override
   public Dimension getMaximumSize() {
