+++ src/view/boardspaces/PropertySpace.java
@@ -1,4 +1,4 @@
-package view.boardSpaces;
+package view.boardspaces;
 
 import java.awt.BorderLayout;
 import java.awt.Color;
