+++ src/view/boardspaces/Spaces.java
@@ -1,4 +1,4 @@
-package view.boardSpaces;
+package view.boardspaces;
 
 
 import Model.GamePackage.Player;
@@ -7,7 +7,7 @@ import java.util.LinkedList;
 import javax.swing.JPanel;
 
 /**********************************************************************
- * The Spaces class is an abstract class which all of the boardSpaces
+ * The Spaces class is an abstract class which all of the boardspaces
  * draw from primarily to draw player tokens on the screen.
  *
  * @author David Baas
